+++ sdk/base-sys/src/bindings/input_note.rs
@@ -3,7 +3,7 @@ use alloc::vec::Vec;
 
 use miden_stdlib_sys::{Felt, Word};
 
-use super::types::{AccountId, Asset, NoteIdx, Recipient};
+use super::types::{AccountId, Asset, NoteIdx, NoteMetadata, Recipient};
 
 #[allow(improper_ctypes)]
 unsafe extern "C" {
@@ -17,7 +17,7 @@ unsafe extern "C" {
     pub fn extern_input_note_get_recipient(note_index: Felt, ptr: *mut Recipient);
 
     #[link_name = "miden::protocol::input_note::get_metadata"]
-    pub fn extern_input_note_get_metadata(note_index: Felt, ptr: *mut Word);
+    pub fn extern_input_note_get_metadata(note_index: Felt, ptr: *mut NoteMetadata);
 
     #[link_name = "miden::protocol::input_note::get_sender"]
     pub fn extern_input_note_get_sender(note_index: Felt, ptr: *mut AccountId);
@@ -82,10 +82,10 @@ pub fn get_recipient(note_index: NoteIdx) -> Recipient {
     }
 }
 
-/// Returns the metadata of the input note at `note_index`.
-pub fn get_metadata(note_index: NoteIdx) -> Word {
+/// Returns the attachment and metadata header of the input note at `note_index`.
+pub fn get_metadata(note_index: NoteIdx) -> NoteMetadata {
     unsafe {
-        let mut ret_area = ::core::mem::MaybeUninit::<Word>::uninit();
+        let mut ret_area = ::core::mem::MaybeUninit::<NoteMetadata>::uninit();
         extern_input_note_get_metadata(note_index.inner, ret_area.as_mut_ptr());
         ret_area.assume_init().reverse()
     }