+++ tests/integration/src/rust_masm_tests/rust_sdk/base/mod.rs
@@ -4,6 +4,5 @@ pub mod account;
 pub mod asset;
 pub mod faucet;
 pub mod input_note;
-pub mod note;
 pub mod output_note;
 pub mod tx;