+++ tests/rust-apps-wasm/rust-sdk/issue-invalid-stack-offset-movup/src/lib.rs
@@ -33,7 +33,7 @@ impl InvalidStackOffsetMovupNote {
         let swapp_note_creator_id = AccountId::new(inputs[8], inputs[9]);
 
         if swapp_note_creator_id == executing_account_id {
-            active_note::add_assets_to_account();
+            // active_note::add_assets_to_account();
             return;
         }
 
@@ -72,7 +72,7 @@ impl InvalidStackOffsetMovupNote {
         let offered_out =
             calculate_output_amount(offered_asset_total, requested_asset_total, input_amount);
 
-        active_note::add_assets_to_account();
+        // active_note::add_assets_to_account();
 
         let routing_serial = add_word(current_note_serial, Word::from_u64_unchecked(0, 0, 0, 1));
 
@@ -189,7 +189,7 @@ fn create_swapp_note(
 
 /// Extracts the note tag from the active note metadata.
 fn get_note_tag() -> Tag {
-    let metadata = active_note::get_metadata();
+    let metadata = active_note::get_metadata().header;
     let left_shifted_32 = metadata[2] * Felt::from_u32(2u32.pow(32));
     let tag_felt = left_shifted_32 / (Felt::from_u32(2u32.pow(32)));
     Tag::from(tag_felt)
@@ -197,7 +197,7 @@ fn get_note_tag() -> Tag {
 
 /// Extracts the note type from the active note metadata.
 fn get_note_type() -> NoteType {
-    let metadata = active_note::get_metadata();
+    let metadata = active_note::get_metadata().header;
     let second_felt = metadata[1];
     let left_shifted_56 = second_felt * Felt::from_u32(2u32.pow(56));
     let note_type_felt = left_shifted_56 / Felt::from_u32(2u32.pow(62));